+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.58.0] - 2023-10-19
+
+### Changed
+
+- Added quote entity to the graphql API
+
 ## [3.57.1] - 2023-10-13
 
 ### Changed